+++ R/Install.R
@@ -3,7 +3,7 @@
 #' @description This is similar to [install.packages()], except it takes by
 #' default the list of packages from `.packages_to_install` in
 #' `SciViews:TempEnv`. That list is populated automatically by infructuous calls
-#' to `package()`, so that just a call to `install()` without arguments is
+#' to `package()`, so that just a call to `Install()` without arguments is
 #' generally sufficient.
 #'
 #' @param pkgs The list of packages to install (character vector). If missing,