+++ tests/testthat.R
@@ -1,3 +1,4 @@
-library("testthat")
+library(testthat)
+library(svMisc)
 
 test_check("svMisc")