+++ NEWS.md
@@ -1,6 +1,22 @@
+# svMisc 1.2.3
+
+- Several examples set as \dontrun{} because time to run all examples is too long on CRAN.
+
+# svMisc 1.2.2
+
+- Link to a JSS paper replaced by the DOI link.
+
+# svMisc 1.2.1
+
+- No more assignation (`last.warning`) into base environment: not permitted any more in R >= 4.1.0.
+
+# svMisc 1.2.0
+
+- Two new functions `pcloud()` and `pcloud_crypto()` ease the management of files that are located in a p-Cloud drive (p-Cloud is a cloud storage system).
+
 # svMisc 1.1.4
 
-- Minor changes to avoid notes in latest devel R CMD check.
+- Minor changes in the documentation to avoid notes in latest devel R CMD check.
 
 # svMisc 1.1.3